requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Werkzeug==3.0.3
snowflake-connector-python==3.10.1
Jinja2==3.1.4
itsdangerous==2.2.0
click==8.1.7
blinker==1.8.2
MarkupSafe==2.1.5
asn1crypto==1.5.1
cffi==1.16.0
cryptography==42.0.7
pyOpenSSL==24.1.0
pyjwt==2.8.0
pytz==2024.1
requests==2.32.2
packaging==24.0
charset-normalizer==3.3.2
idna==3.7
certifi==2024.2.2
typing-extensions==4.12.0
filelock==3.14.0
sortedcontainers==2.4.0
platformdirs==4.2.2
tomlkit==0.12.5
pycparser==2.22
urllib3==2.2.1